% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_sdm_rasters_threshold.R
\name{conv_sdm_rasters_threshold}
\alias{conv_sdm_rasters_threshold}
\title{Convert maxent output rasters to binary by threshold values}
\usage{
conv_sdm_rasters_threshold(raster_path, results_file, threshold)
}
\arguments{
\item{raster_path}{path to the .asc file}

\item{results_file}{path to the results file}

\item{threshold}{either "balanced", "10pct", or "minimum"}
}
\description{
Convert maxent output rasters to binary by threshold values
}
