% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_function.R
\name{sdm_function}
\alias{sdm_function}
\title{sdm_function}
\usage{
sdm_function(directory.path, coords = NULL, species.name,
  climate.map.path = NULL, threshold = "balanced")
}
\arguments{
\item{directory.path}{path to an empty directory to store data and maxent outputs}

\item{coords}{a vector of four boundary coordinates: min lon, max lon, min lat, max lat}

\item{species.name}{the scientific name of the species to be mapped}

\item{climate.map.path}{map to a global climate map. The path should be to the folder containing 19 raster layers. If NA, the map will be downloaded to directory.path.}

\item{threshold}{Default is "balanced," but other options include "10pct" and "minimum"}
}
\description{
Gets map of species distribution from maxent model
}
